+++ macos/Classes/AudioSessionPlugin.m
@@ -17,10 +17,10 @@ static NSMutableArray<AudioSessionPlugin *> *plugins = nil;
 - (instancetype)initWithRegistrar:(NSObject<FlutterPluginRegistrar> *)registrar {
     self = [super init];
     NSAssert(self, @"super init cannot be nil");
-	_channel = [FlutterMethodChannel
-		methodChannelWithName:@"com.ryanheise.audio_session"
+    _channel = [FlutterMethodChannel
+        methodChannelWithName:@"com.ryanheise.audio_session"
               binaryMessenger:[registrar messenger]];
-	[registrar addMethodCallDelegate:self channel:_channel];
+    [registrar addMethodCallDelegate:self channel:_channel];
     [plugins addObject:self];
     return self;
 }
@@ -31,17 +31,17 @@ static NSMutableArray<AudioSessionPlugin *> *plugins = nil;
 
 - (void)handleMethodCall:(FlutterMethodCall*)call result:(FlutterResult)result {
     NSArray* args = (NSArray*)call.arguments;
-	if ([@"setConfiguration" isEqualToString:call.method]) {
+    if ([@"setConfiguration" isEqualToString:call.method]) {
         configuration = args[0];
         for (int i = 0; i < plugins.count; i++) {
             [plugins[i].channel invokeMethod:@"onConfigurationChanged" arguments:@[configuration]];
         }
-		result(nil);
-	} else if ([@"getConfiguration" isEqualToString:call.method]) {
+        result(nil);
+    } else if ([@"getConfiguration" isEqualToString:call.method]) {
         result(configuration);
-	} else {
-		result(FlutterMethodNotImplemented);
-	}
+    } else {
+        result(FlutterMethodNotImplemented);
+    }
 }
 
 - (void) dealloc {