+++ example/ios/RunnerTests/RunnerTests.m
@@ -1,0 +1,36 @@
+#import <Flutter/Flutter.h>
+#import <UIKit/UIKit.h>
+#import <XCTest/XCTest.h>
+
+// If your plugin has been explicitly set to "type: .dynamic" in the Package.swift,
+// you will need to add your plugin as a dependency of RunnerTests within Xcode.
+
+@import audio_session;
+
+// This demonstrates a simple unit test of the Objective-C portion of this plugin's implementation.
+//
+// See https://developer.apple.com/documentation/xctest for more information about using XCTest.
+
+@interface RunnerTests : XCTestCase
+
+@end
+
+@implementation RunnerTests
+
+- (void)testExample {
+  AudioSessionPlugin *plugin = [[AudioSessionPlugin alloc] init];
+
+  FlutterMethodCall *call = [FlutterMethodCall methodCallWithMethodName:@"getPlatformVersion"
+                                                              arguments:nil];
+  XCTestExpectation *expectation = [self expectationWithDescription:@"result block must be called"];
+  [plugin handleMethodCall:call
+                    result:^(id result) {
+                      NSString *expected = [NSString
+                          stringWithFormat:@"iOS %@", UIDevice.currentDevice.systemVersion];
+                      XCTAssertEqualObjects(result, expected);
+                      [expectation fulfill];
+                    }];
+  [self waitForExpectationsWithTimeout:1 handler:nil];
+}
+
+@end