+++ example/test/widget_test.dart
@@ -5,7 +5,7 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
-//import 'package:flutter/material.dart';
+// import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:audio_session_example/main.dart';
@@ -13,15 +13,15 @@ import 'package:audio_session_example/main.dart';
 void main() {
   testWidgets('Verify Platform version', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(MyApp());
+    await tester.pumpWidget(const MyApp());
 
-    //// Verify that platform version is retrieved.
-    //expect(
-    //  find.byWidgetPredicate(
-    //    (Widget widget) => widget is Text &&
-    //                       widget.data!.startsWith('Running on:'),
-    //  ),
-    //  findsOneWidget,
-    //);
+    // Verify that platform version is retrieved.
+    // expect(
+    //   find.byWidgetPredicate(
+    //     (Widget widget) => widget is Text &&
+    //                        widget.data!.startsWith('Running on:'),
+    //   ),
+    //   findsOneWidget,
+    // );
   });
 }